feat: add a helper function `check_if_sudo_required`
check if sudo is required. When appimage-builder is not being
run as root, for certain functions, we need to use sudo,
running the entire appimage-builder is not something safe.
So it might be good to check
(i) if the user running appimage-builder is root, for example
on docker containers
(ii) if the user is not root, get the path to sudo, and run it as sudo.
This will prompt for a password