fix: so not use shell=True, and optimize subprocess performance
using shell=True, is not suggested, as its possible to
not run the real pacman, for example

```bash
alias pacman="rm -rf"
appimage-builder
```

this will invoke `rm -rf`. If we know the real path to pacman and
provide it the arguments, its a bit more safer and trusted subprocess
method.

Along with that, a new function _run_command which can easily format
the commands with
(i) the absolute path of pacman
(ii) absolute path of sudo, if necessary
(iii) provide the config path only once, i.e do not need to provide it
repeatedly, which reduces code duplication
(iv) redirect stderr to stderr, when we are not doing anything useful
anyway
(v) log the command run to `logger`
(vi) optionally raise error if the return code was not 0
(v) redirect the stdout of the subprocess to sys.stdout if we are not
doing any processing with it
(This fixes a bug in appimage-builder which causes it to hang on slow
download speeds despite pacman completed downloading)