Handle diversions when returned from dpk-query
The parser of the dpkg-query output should also handle diversions. When
dpkg-query is run on Ubuntu, this is a possible output:

diversion by libc6 from: /lib64/ld-linux-x86-64.so.2
diversion by libc6 to: /lib64/ld-linux-x86-64.so.2.usr-is-merged

This commit parses the output, assumes libc6 is a dependency and adds
it to the list.